+++ mirrord-layer/src/file.rs
@@ -51,6 +51,7 @@ static IGNORE_FILES: LazyLock<RegexSet> = LazyLock::new(|| {
         // TODO: `node` searches for this file in multiple directories, bypassing some of our
         // ignore regexes, maybe other "project runners" will do the same.
         r".*/package.json",
+        r".*/.cache/deno/.*",
         &current_dir.to_string_lossy(),
     ])
     .unwrap();